requirements.txt
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.7
pylint==1.8.4
python-twitter==3.4.1
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
